update model version 1.0.3
Change cost parameter beta and gamma powers. Removing the minus.

after discussion:

- For the risk to an individual: 1-r is the proportion (again between 0
and 1) of the rhinos that are not devalued, we use it to correlate with
risk as if the manager devalues a lot of rhinos they will not be able to
spend money on security. Thus the risk to the individual should increase
as 1 - r increase. So the utility should be $(1-r) ^ {beta}$.

- For the cost to a poacher (1-rs) is the cost associated with finding a
rhino. As this increases the overall cost increases, thus it should be:
$(1-rs)^{\gamma}$.

- But the poachers are exposed to an extra cost as well, whether they are being
selective or not. Here the information of the proportion of rhinos that have
been devalued is not known to the poachers. The argument is that a selective poacher will take his/her time
and not kill instantly and that \textbf{should be more expensive}. This cost
should be increasing as the value of $r$ is increases.