First draft of the rhino work.
The structure is the following:

- Abstract
- Introduction
- The model
- The evolutionary model
- Numerical analysis
- Discussion and conclusion

Several things stand up.*(Please feel free to go through my maths)*

It can be proven analytically that selective is never stable and
on the other hand indiscriminate is always stable. There seem
to be a few mixed strategies that are stable but are very close
to 0.

The only left to study through numerical experiments is the
evolutionary stability. None of the strategies seem to be ESS.

The conclusion I have made is that poachers will chose to
kill as long as they believe the is even the slightest gain.

- I did re run the full analysis
- Evolutionary approach from Vince's note have been implemented.
We need to take a look at them together because I am not entirly sure
how to interpret them