draft 20.09
This is the draft after the meeting where everything
is now proven analytically.

I have fixed the introdction as described. But I have not
commit before this. Thus in this draft the introduction
is fixed but everything remains as in the old version that
I had. Comming now just to keep the history.