minor corrections
- addressed mike small comments
- standarised references in bibliography
- added mike and david in the acknowledgments

Addressing Mike's comments include the first sentence of
the abstract. I did change the opening sentence and
added a similar sentence in the introduction with a reference.

Another comment from Mike was that the
`New rhino conservation project in ...' reference was old (1992).
I added an extra one without removing the 1992 reference.

Added markedu-up.pdf