Restructed to fit american naturalist journal. (#29)
* restrured to fit american naturalist journal.

This is the version for submission to the american naturalist
journal(http://www.journals.uchicago.edu/journals/an/instruct).

Although the reviewing the journal does not need any spefici formatting
there are some guidelines and an example latex file to make the life
of the reviewers easier. The guidelines include:

- The first page of the manuscript file should include,
  - the title
  - four-six keywords
  - the number of words in the text
  - a list of all the elements of the manuscript that will appear in the expanded online edition
  - type of manyscript

- For the reviewing process to be objective,
  - all authors name and contact infos were commented out
  - acknowledgments as well
  - link to Github accounts must be paste in the comments field during submission

- The guideline come with a latex template example. I tried to follow the template,
  - fix bibliography to fit the journal
  - abstract as a section
  - new page after abstract
  - sections are not enumerated thus refer to sections by name.

* create markdown for submission.

* cover letter attempt and minor corrections.

Minor corrections include,
- removing the list of gigures from the first page
- fixing citation style.

Also my attempt for a cover letter. I struggled with it.

* address vince's suggestions

* fix references.

Archived data to zenodo and added to references. Also standarized
the rest of the references a bit